Refactor: extract operation cancel functions struct to separate file
